Restructure tests to improve PR testing
