+++ test/package.json
@@ -4,7 +4,9 @@
   "description": "Test material for Codewind",
   "main": "index.js",
   "scripts": {
-    "test": "scripts/mocha.sh",
+    "test": "scripts/mocha.sh src/API",
+    "integrationtest": "scripts/mocha.sh src/integration",
+    "unittest": "scripts/mocha.sh src/unit",
     "eslint": "eslint src/**/*.js"
   },
   "bin": {
