+++ test/src/API/projectTypes.test.js
@@ -9,15 +9,29 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 const chai = require('chai');
-const reqService = require('../modules/request.service');
+
+const reqService = require('../../modules/request.service');
+const { ADMIN_COOKIE } = require('../../config');
 
 chai.should();
 
-describe('Root endpoint test', function() {
+describe('Project Types API tests', function() {
 
-    it('should return status 200', async function() {
+    it('should return expected list of project types', async function() {
         const res = await reqService.chai
-            .get('/');
+            .get('/api/v1/project-types')
+            .set('Cookie', ADMIN_COOKIE);
+
         res.should.have.status(200);
+        res.should.have.ownProperty('body');
+        res.body.should.be.an('array');
+        res.body.should.have.members([
+            'liberty',
+            'nodejs',
+            'spring',
+            'swift',
+            'docker',
+            'appsodyExtension',
+        ]);
     });
 });
