+++ test/src/API/projects/build.test.js
@@ -10,8 +10,8 @@
  *******************************************************************************/
 const chai = require('chai');
 
-const projectService = require('../../modules/project.service');
-const { testTimeout } = require('../../config');
+const projectService = require('../../../modules/project.service');
+const { testTimeout } = require('../../../config');
 
 chai.should();
