+++ test/src/API/projects/invalid-name-chars.test.js
@@ -11,8 +11,8 @@
 
 const chai = require('chai');
 
-const reqService = require('../../modules/request.service');
-const { testTimeout, ADMIN_COOKIE } = require('../../config');
+const reqService = require('../../../modules/request.service');
+const { testTimeout, ADMIN_COOKIE } = require('../../../config');
 
 chai.should();
