+++ test/src/API/security/sanitize.test.js
@@ -10,8 +10,8 @@
  *******************************************************************************/
 const chai = require('chai');
 
-const reqService = require('../../modules/request.service');
-const { ADMIN_COOKIE, testTimeout } = require('../../config');
+const reqService = require('../../../modules/request.service');
+const { ADMIN_COOKIE, testTimeout } = require('../../../config');
 
 chai.should();
