+++ test/src/API/templates.test.js
@@ -21,8 +21,8 @@ const {
     deleteTemplateRepo,
     resetTemplateReposTo,
     getTemplateStyles,
-} = require('../modules/template.service');
-const { pathToApiSpec } = require('../config');
+} = require('../../modules/template.service');
+const { pathToApiSpec } = require('../../config');
 
 chai.should();
 chai.use(chaiResValidator(pathToApiSpec));
