+++ test/src/API/templates/enableRepo.test.js
